Add macros for different layout components
This PR add three new macros files for various site components. For this first phase,
code is output from the rails app. Whenever possible, I did my best to break out pieces
into smaller components, but there is still a lot of cleanup and refactoring. This is really a first step at getting the Ando frame work to render pieces of the site.